+++ terraform/main.tf
@@ -14,6 +14,10 @@ terraform {
       source  = "hashicorp/awscc"
       version = "~> 1.0"
     }
+    time = {
+      source  = "hashicorp/time"
+      version = "~> 0.13"
+    }
   }
   backend "s3" {}
 }