+++ terraform/modules/domain/versions.tf
@@ -11,5 +11,9 @@ terraform {
       version               = "~> 6.0"
       configuration_aliases = [aws.us_east_1]
     }
+    time = {
+      source  = "hashicorp/time"
+      version = "~> 0.13"
+    }
   }
 }